+++ plugin/gcp_config.go
@@ -6,7 +6,6 @@ import (
 	"fmt"
 	"strings"
 
-	"github.com/hashicorp/errwrap"
 	"github.com/hashicorp/go-gcp-common/gcputil"
 	"github.com/hashicorp/vault/sdk/framework"
 	"github.com/hashicorp/vault/sdk/helper/authmetadata"
@@ -21,6 +20,15 @@ type gcpConfig struct {
 	IAMAuthMetadata *authmetadata.Handler   `json:"iam_auth_metadata_handler"`
 	GCEAliasType    string                  `json:"gce_alias"`
 	GCEAuthMetadata *authmetadata.Handler   `json:"gce_auth_metadata_handler"`
+
+	// APICustomEndpoint overrides the service endpoint for www.googleapis.com
+	APICustomEndpoint string `json:"api_custom_endpoint"`
+	// IAMCustomEndpoint overrides the service endpoint for api.googleapis.com
+	IAMCustomEndpoint string `json:"iam_custom_endpoint"`
+	// CRMCustomEndpoint overrides the service endpoint for cloudresourcemananger.googleapis.com
+	CRMCustomEndpoint string `json:"crm_custom_endpoint"`
+	// ComputeCustomEndpoint overrides the service endpoint for compute.googleapis.com
+	ComputeCustomEndpoint string `json:"compute_custom_endpoint"`
 }
 
 // standardizedCreds wraps gcputil.GcpCredentials with a type to allow
@@ -55,7 +63,7 @@ func (c *gcpConfig) Update(d *framework.FieldData) error {
 	if v, ok := d.GetOk("credentials"); ok {
 		creds, err := gcputil.Credentials(v.(string))
 		if err != nil {
-			return errwrap.Wrapf("failed to read credentials: {{err}}", err)
+			return fmt.Errorf("failed to read credentials: %w", err)
 		}
 
 		if len(creds.PrivateKeyId) == 0 {
@@ -67,25 +75,40 @@ func (c *gcpConfig) Update(d *framework.FieldData) error {
 
 	rawIamAlias, exists := d.GetOk("iam_alias")
 	if exists {
-		iamAlias := rawIamAlias.(string)
-		if iamAlias != c.IAMAliasType {
-			c.IAMAliasType = iamAlias
-		}
+		c.IAMAliasType = rawIamAlias.(string)
 	}
+
 	if err := c.IAMAuthMetadata.ParseAuthMetadata(d); err != nil {
-		return errwrap.Wrapf("failed to parse iam metadata: {{err}}", err)
+		return fmt.Errorf("failed to parse iam metadata: %w", err)
 	}
 
 	rawGceAlias, exists := d.GetOk("gce_alias")
 	if exists {
-		gceAlias := rawGceAlias.(string)
-		if gceAlias != c.GCEAliasType {
-			c.GCEAliasType = gceAlias
+		c.GCEAliasType = rawGceAlias.(string)
+	}
+
+	if err := c.GCEAuthMetadata.ParseAuthMetadata(d); err != nil {
+		return fmt.Errorf("failed to parse gce metadata: %w", err)
+	}
+
+	rawEndpoint, exists := d.GetOk("custom_endpoint")
+	if exists {
+		for k, v := range rawEndpoint.(map[string]string) {
+			switch k {
+			case "api":
+				c.APICustomEndpoint = v
+			case "iam":
+				c.IAMCustomEndpoint = v
+			case "crm":
+				c.CRMCustomEndpoint = v
+			case "compute":
+				c.ComputeCustomEndpoint = v
+			default:
+				return fmt.Errorf("invalid custom endpoint type %q. Available types are: 'api', 'iam', 'crm', 'compute'", k)
+			}
 		}
 	}
-	if err := c.GCEAuthMetadata.ParseAuthMetadata(d); err != nil {
-		return errwrap.Wrapf("failed to parse gce metadata: {{err}}", err)
-	}
+
 	return nil
 }